Implement routing for the testit-daemon
Implements the routing for the testit-daemon. This is the first step
towards the implementation of the testit-daemon. The routing is
implemented using the reqwest client. The routing supports
All http versions, all methods, and all headers.  It allows being
proxied to both http and https endpoints however it is missing a
lot of functionality to make it fully functional.

The routing is implemented in the server.rs file. The server.rs file
is responsible for handling the incoming requests and routing them
to the appropriate endpoint. The server.rs file is also responsible
for handling the responses from the endpoint and returning them to
the client.

Future improvements
- TLS support
- Better error handling
- Better logging
- Improved configuration
- Improved testing

Resolves: None
See also: None