Added functionality upgrades
Added support for TLS12 or TLS13 for the server configuration. The
enum allows for TLS11 and TLS10, but rusttls does not seem to support
these versions. If TLS11 or TLS10 are selected, the server will
default to TLS12.

The following configuration has been added to the routing configuration:
* verbose, default false
* http1_only, default false
* accept_invalid_certs, default false
* accept_invalid_hostnames, default false
* min_tls_version, default TLS10
* max_tls_version, default TLS13
* read_timeout, using reqwest default
* connect_timeout, using reqwest default

Resolves: None
See also: None