Add tcp listener
Adds a new tcp listener to the daemon. This is a new feature that
allows the daemon to listen for tcp connections and respond with data
from a file or directly from the config. This is useful for testing tcp
clients.

Currently the listener will read until the buffer is empty, then wait
a specified time and then write the response. The listener will allways
close the connection after the response is sent.

Future improvements
- Keep the connection open and respond to multiple requests
- Look for a way to drop connections
- Wait for specific data before responding

Resolves: None
See also: None