refactor: Codebase refactored to improve readability
Previously the code was not very readable and had a lot of duplication.
This commit refactors the code to make it more readable and remove
duplication. There are stil some duplication remaining which will be
addressed in future commits.

The code is now more maintanable as the code has been modularized and
each module is responsible for a single task. This can still be improved
by further modularizing the code.

Tauri has been improved to use a more standardized api and the code has
been refactored to use the new api. This will make it easier to
move the code into their own components in the future.

Breaking changes: A few bugs have been introduced in the refactoring
process. These will be fixed in the next commit. The json files have
been slightly changed by using camelcase instead of snake_case.

Resolves: None