chore: Reintroduce outdated to build process (#6)
* chore: Reintroduce outdated to build process

This commit reintroduces the outdated crate to the build process. This
is necessary to ensure that the build process is able to detect outdated
dependencies and alert the developer to the need to update them.

The outdated crate is now running on each push to a branch, and
only checks root dependencies.

Future improvements: None

Breaking changes: None

Resolves: #3