feat/editable http headers (#11)
* feat: Editable http headers

This feature allows users to edit the http headers of a mock response
in the UI. The headers is editable in a textarea, and the user can
input the headers in the format of key-value pairs separated by a colon
and a newline. Unparseable headers will be ignored.

Future improvements: Better validation of headers, and a more
user-friendly way to edit headers.

Breaking changes: None

Resolves: #5