feat: Settings component (#14)
* feat: Settings component

Adds a settings component to the UI that allows the user to change the
settings of the application. The settings are stored in a JSON file in
the user's home directory. The settings are loaded when the application
starts and saved when the user changes them. The settings are also
loaded when a test is started.

Currently only the setting for what daemon to use in tests are
supported. The settings component is designed to be easily extensible
to support more settings in the future.

Also upgrades tokio dependency.

Future improvements: More settings

Breaking changes: None

Resolves: #4