chore: Refactor tcp and http servers
The server module is now a module that contains the tcp and http
modules including the setup. They also now have a common module that
contains the common trait that all servers must implement. This will
make it easier to add more servers in the future.

It code now also uses a vector of servers instead of a vector for
each server type.

Future improvements: Move common functionality to the common module.-

Breaking changes: None

Resolves: #118