Chore: refactor tcp http servers (#19)
* feat: editable-http-headers (#9)

feat: Editable http headers

This feature allows users to edit the http headers of a mock response
in the UI. The headers is editable in a textarea, and the user can
input the headers in the format of key-value pairs separated by a colon
and a newline. Unparseable headers will be ignored.

Future improvements: Better validation of headers, and a more
user-friendly way to edit headers.

Breaking changes: None

Resolves: #5

* Revert "feat: editable-http-headers (#9)" (#10)

This reverts commit 87598d4897bdafd5775246314916216fa4554976.

* chore: Reintroduce outdated to build process (#6)

* chore: Reintroduce outdated to build process

This commit reintroduces the outdated crate to the build process. This
is necessary to ensure that the build process is able to detect outdated
dependencies and alert the developer to the need to update them.

The outdated crate is now running on each push to a branch, and
only checks root dependencies.

Future improvements: None

Breaking changes: None

Resolves: #3

* chore: Reintroduce clippy to build process (#7)

* chore: Reintroduce clippy to build process

Reintroduce clippy to the build process. This will help us to catch
potential issues and improvements in the codebase. The test will fail if
clippy detects any issues.

This commit also includes some minor improvements in the codebase to
resolve clippy warnings.

Future improvements: None

Breaking changes: None

Resolves: #2

* chore: Refactor tcp and http servers

The server module is now a module that contains the tcp and http
modules including the setup. They also now have a common module that
contains the common trait that all servers must implement. This will
make it easier to add more servers in the future.

It code now also uses a vector of servers instead of a vector for
each server type.

Future improvements: Move common functionality to the common module.-

Breaking changes: None

Resolves: #18