feat: Improved validation for files (#20)
Adds file validation to the API, so that only files with the correct
extension is used and that they exist. This is done using the Tauri
dialog api. The input fields are now read only and selected files are
displayed. The user can remove files by clicking the clear button.

Future improvements:
- Show a warning if the file no longer exist.
- Show a warning if both file and data is present for tcp listeners.

Breaking changes: None

Resolves: #17