fix: fix endpoint initialization (#28)
Fixes the endpoint initialization so that when the endpoint is added
to the configuration, it is initalized in an ok state. This is done
by adding ok values when initializing the endpoint.

This also fixes an issue in the Test.vue component when the delay
is set to 0. This caused the validation to fail as the value was a
number and not a string. The value is now success when retrieved as a
number.

Future improvements: None

Breaking changes: None

Resolves: #27