feat: support input parameters in mock responses
Allows for input parameters in mock responses, status codes, and
headers. This is done by adding ${keyname} to be converted with
the value of the keyname in the request.

Future improvements: None

Breaking changes: Mock response code is now a string instead of
an integer. This allows for more flexibility in the response code.

Resolves: #55