feat: support input parameters (#55)
* feat: support input parameters

Adds support for input parameters in the API. This allows users to
pass parameters to the mock returns so that they can be reused
across different requests. The parameters are defined in the
configuration file and can be used in the mock returns.

Allows for input parameters in mock responses, status codes, and
headers. This is done by adding ${keyname} to be converted with
the value of the key name in the request.

Adds support for input parameters in the UI. Allows to add and remove
input parameters in the test dialog. There is no validation of the
parameters.

Fixes multiple clippy issues and formats.

Future improvements: None

Breaking changes: Mock response code is now a string instead of
an integer. This allows for more flexibility in the response code.

Resolves: #54