feat: sort params (#61)
Changes the params element to be a sorted array in the model. The
sorting occurs when converting the config model to the ui model. This
is done to make reading the params easier in the UI. The sorting is
done by alphabethic order of the param name. The sorting is done in the
model.rs file.

Changes the ui for the params so they are easier to read.

Future improvements: None

Breaking changes: None

Resolves: #60