feat: update test data on test view
This feature allows users to update the test data directly from the
test view. This is the same as the functionality in the tests data view.
This is useful for users who want to quickly update the test data
without having to navigate to the tests data view.

Allows updating
- name
- description
- parameters

Future improvements: Reusable component so it's not duplicated.

Breaking changes: None

Resolves: #62