feat: add predefined parameter sets (#69)
* feat: add predefined parameter sets

Allows users to define parameter sets in the config file, which can be
used to quickly switch between different configurations. The predefined
parameter sets are defined in the config file and can be selected
using the `--predefined_set=<name>` flag.

Predefined parameters can be used in conjunction  with --param flags.
The --param flags will not override the parameters defined in the
predefined set.

Future improvements: Add override functionality for --param flags to
override the parameters defined in the predefined set.

Breaking changes: None

Resolves: #64