fix: multiple errors shown if refresh fails (#78)
Fixes the issue that multiple errors are shown if the refresh fails.
This was caused by each of the refresh calls in the `refresh` method of
the `Test.vue` component. Now only the get test call shows error
dialog if error occurs, and the other refresh calls only log to
the console.

This is a bug fix that improves the user experience by preventing
multiple error dialogs from appearing when the refresh fails.

Future improvements: None

Breaking changes: None

Resolves: #77