feat: set parameters when running apinae UI
Allow users to set parameters when running apinae through the UI. When
the user runs apinae through the UI and the test requires parameters,
the UI will prompt the user to enter the parameters. The parameters will
be passed to apinae when it is run.

Future improvements: Allow choosing preset parameters. Close parameter
dialog when run. Remember last used parameters.

Breaking changes: None

Resolves: #63