feat: add delay to routes (#83)
* feat: add delay to routes

Adds delay before request and request after response to routes.
This allows for testing of the delay features. The ui has been
updated to allow for setting the delays in the route editor.
The delay is set in milliseconds and can be set to 0 or none for
no delay.

Future improvements: None

Breaking changes: None

Resolves: #82