feat: rename tests to setups (#87)
This commit renames the "tests" to "setups" in the codebase to better
reflect their purpose. The term "tests" was misleading, as these are not
actual tests but rather preparations and setups for the tests to be run.

This is a major refactor and some parts might be missing or broken.

Future improvements: None

Breaking changes: The tests array in the config file has been renamed
to setups. This will require manual changes in the config file. The old
tests array will be ignored and the setups array will be used instead.

Resolves: #84