feat: Fixes and documentantation improvements
Fixes a bug in the http server when checking for valid endpoints. Now
when regexp are included, but body is not present it's an invalid
endpoint. Adds multiple tests for the validation.

Adds improved documentation for multiple methods.

Adds regexp documentation example in the docs. This will be improved
in the future.

Fixes a bug in the UI where the mock delay validation failed initially
due to a missing check. Now validationOptionalNumber checks for typeof
number and if successful then returns is-valid.

Future improvements: None

Breaking changes: None

Resolves: #101,#102, #103, #104