feat: UI improvements
UI improvements have been made to the setup view for the endpoints.
The layout has been adjusted to provide a more user-friendly experience,
by making endpoints into accordions with id, path, method and body
expressions on the accordion and the rest in the body.

The endpoint accordion now has a better structure for larger and smaller
screens.

Future improvements: None

Breaking changes: None

Resolves: #105