build: improve-build-scripts (#112)
Updates release so that version is retrived from the tag. It installs
cargo-edit where we use set-version. Version set is always the tag
version with v removed.

Fixes the tauri build due to a change in latest tauri-cli. Now prefix
is added so it no longer fails on npm run build.

Breaking changes: None

Resolves: #111